+++ internal/taskfile/taskfile.go
@@ -1,5 +1,10 @@
 package taskfile
 
+import (
+	"fmt"
+	"strconv"
+)
+
 // Taskfile represents a Taskfile.yml
 type Taskfile struct {
 	Version    string
@@ -7,8 +12,8 @@ type Taskfile struct {
 	Output     string
 	Method     string
 	Includes   IncludedTaskfiles
-	Vars       Vars
-	Env        Vars
+	Vars       *Vars
+	Env        *Vars
 	Tasks      Tasks
 	Silent     bool
 }
@@ -21,8 +26,8 @@ func (tf *Taskfile) UnmarshalYAML(unmarshal func(interface{}) error) error {
 		Output     string
 		Method     string
 		Includes   IncludedTaskfiles
-		Vars       Vars
-		Env        Vars
+		Vars       *Vars
+		Env        *Vars
 		Tasks      Tasks
 		Silent     bool
 	}
@@ -42,7 +47,19 @@ func (tf *Taskfile) UnmarshalYAML(unmarshal func(interface{}) error) error {
 		tf.Expansions = 2
 	}
 	if tf.Vars == nil {
-		tf.Vars = make(Vars)
+		tf.Vars = &Vars{}
+	}
+	if tf.Env == nil {
+		tf.Env = &Vars{}
 	}
 	return nil
+}
+
+// ParsedVersion returns the version as a float64
+func (tf *Taskfile) ParsedVersion() (float64, error) {
+	v, err := strconv.ParseFloat(tf.Version, 64)
+	if err != nil {
+		return 0, fmt.Errorf(`task: Could not parse taskfile version "%s": %v`, tf.Version, err)
+	}
+	return v, nil
 }