+++ internal/templater/templater.go
@@ -2,6 +2,7 @@ package templater
 
 import (
 	"bytes"
+	"strings"
 	"text/template"
 
 	"github.com/go-task/task/v2/internal/taskfile"
@@ -12,7 +13,8 @@ import (
 // happen will be assigned to r.err, and consecutive calls to funcs will just
 // return the zero value.
 type Templater struct {
-	Vars taskfile.Vars
+	Vars          *taskfile.Vars
+	RemoveNoValue bool
 
 	cacheMap map[string]interface{}
 	err      error
@@ -42,6 +44,9 @@ func (r *Templater) Replace(str string) string {
 		r.err = err
 		return ""
 	}
+	if r.RemoveNoValue {
+		return strings.ReplaceAll(b.String(), "<no value>", "")
+	}
 	return b.String()
 }
 
@@ -57,20 +62,22 @@ func (r *Templater) ReplaceSlice(strs []string) []string {
 	return new
 }
 
-func (r *Templater) ReplaceVars(vars taskfile.Vars) taskfile.Vars {
-	if r.err != nil || len(vars) == 0 {
+func (r *Templater) ReplaceVars(vars *taskfile.Vars) *taskfile.Vars {
+	if r.err != nil || vars == nil || len(vars.Keys) == 0 {
 		return nil
 	}
 
-	new := make(taskfile.Vars, len(vars))
-	for k, v := range vars {
-		new[k] = taskfile.Var{
+	var new taskfile.Vars
+	vars.Range(func(k string, v taskfile.Var) error {
+		new.Set(k, taskfile.Var{
 			Static: r.Replace(v.Static),
 			Live:   v.Live,
 			Sh:     r.Replace(v.Sh),
-		}
-	}
-	return new
+		})
+		return nil
+	})
+
+	return &new
 }
 
 func (r *Templater) Err() error {